Added delta time for updates

+++ src/Entity.h
@@ -32,7 +32,7 @@ class Entity
 public:
 	Entity(Vector2 position = { 0, 0 }) : m_Position(position){}
 
-	virtual void OnUpdate(float delta) {}
+	virtual void OnUpdate(double delta) {}
 	virtual void OnDraw() {}
 protected:
 	Vector2 m_Position;
